requirements.txt
# === 核心交易 API ===
shioaji>=1.2.0

# === Web 框架 ===
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0

# === 資料庫 ===
sqlalchemy>=2.0.0

# === 回測框架 ===
backtrader>=1.9.78

# === 通知 ===
python-telegram-bot>=21.0

# === 數據分析 ===
numpy>=1.24.0
pandas>=2.0.0

# === 設定管理 ===
pyyaml>=6.0
python-dotenv>=1.0.0

# === 技術指標 ===
ta>=0.11.0
